data-collection-system/requirements.txt
requests
psycopg2-binary
python-dotenv
pyfaktory

